Adding test application for automated testing.
Scala compatibility is limited in play-authenticate currently, but not
entirely absent. Automated testing should improve this situation.

This test application provides a very basic automated testing
environment, which implements a UsernamePasswordAuthProvider and
associated support classes in Java, and tests controller protection for
Java & Scala.

The main aim is to provide very basic test coverage of core
play-authenticate functionality.

Tests of the Scala controller are written in Scala too.


Former-commit-id: 97504cfc30f5572a995d761a33ed9fb8438d811f